refactor(palette): split App.tsx under the 500-line monolith cap
App.tsx was 1062 effective lines, over the enforced cap. Extract focused
modules, each well under 500, preserving behavior verbatim:

- components/palette/AxonMark.tsx        — brand SVG
- components/palette/ActionIcon.tsx      — icon + subcommand→icon map
- components/palette/HistoryPanel.tsx    — recent-runs panel + HistoryItem
- components/palette/ActionList.tsx      — searchable action list (JSX)
- components/palette/PaletteFooter.tsx   — footer hints + status (JSX)
- lib/appHelpers.ts                      — pure helpers
- lib/useCrawlJob.ts                     — live crawl poll + job handlers
- lib/useActionRunner.ts                 — submit + ask-stream effect + history
- lib/invoke.ts                          — now also exports shared appWindow

App.tsx keeps the shared state and render composition (475 effective lines).
The two hooks receive App-owned run/history/panel setters (no useCallback,
so closures match the originals). submit's dead `= active` default dropped;
every call site already passes the action explicitly.

Verified: tsc --noEmit clean; vitest 28/28; monolith --file passes on
App.tsx and every new module; live crawl view exercised end-to-end in the
vite dev harness against a real backend (submit→job→poll snapshot updates,
minimize/expand/view-partial/cancel), zero console errors/exceptions.

Closes axon_rust-wdqi
